fix: Se arregló la implementación incorrecta del código

+++ src/main/java/com/vegas/sistema_gestion_operativa/sales/application/dto/ProductSalesStatsDto.java
@@ -3,10 +3,12 @@ package com.vegas.sistema_gestion_operativa.sales.application.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ProductSalesStatsDto {
     private Long productId;
     private String productName;
-    private Long totalQuantitySold;
+    private BigDecimal totalQuantitySold;
 }
